pyproject.toml
[project]
name = "mediawiki-api-mcp"
version = "0.1.0"
description = "Model Context Protocol server for MediaWiki API integration"
authors = [
    {name = "Peter J. Herrel", email = "peterherrel@protonmail.com"}
]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest-asyncio>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"

[project.scripts]
mediawiki-api-mcp = "mediawiki_api_mcp.server:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
